ice-breaker/pyproject.toml
[project]
name = "ice-breaker"
version = "0.0.1"
description = "Health Professional Ice Breaker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "firecrawl-py>=2.5.3",
    "langchain>=0.3.24",
    "langchain-aws>=0.2.22",
    "langchain-community>=0.3.23",
    "langchain-tavily>=0.1.6",
    "langgraph>=0.4.1",
    "langsmith>=0.3.39",
    "loguru>=0.7.3",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
]
